+++ Riot/Modules/SecureBackup/Setup/Intro/SecureBackupSetupIntroViewModel.swift
@@ -15,17 +15,19 @@
 //
 
 import Foundation
-import MatrixKit
 
-struct RoomInfoBasicViewData {
-    let avatarUrl: String?
-    let mediaManager: MXMediaManager?
+final class SecureBackupSetupIntroViewModel: SecureBackupSetupIntroViewModelType {
     
-    let roomId: String
-    let roomDisplayName: String?
-    let mainRoomAlias: String?
-    let roomTopic: String?
-    let encryptionImage: UIImage?
-    let isEncrypted: Bool
-    let isDirect: Bool
+    // MARK: - Properties
+
+    // TODO: Make these properties private
+    let keyBackup: MXKeyBackup?
+    let checkKeyBackup: Bool
+    
+    // MARK: - Setup
+    
+    init(keyBackup: MXKeyBackup?, checkKeyBackup: Bool) {
+        self.keyBackup = keyBackup
+        self.checkKeyBackup = checkKeyBackup
+    }    
 }