+++ Riot/Modules/SecureBackup/Setup/SecureBackupSetupCoordinator.swift
@@ -29,6 +29,7 @@ final class SecureBackupSetupCoordinator: SecureBackupSetupCoordinatorType {
     private let session: MXSession
     private let recoveryService: MXRecoveryService
     private let keyBackup: MXKeyBackup?
+    private let checkKeyBackup: Bool
 
     // MARK: Public
 
@@ -39,10 +40,16 @@ final class SecureBackupSetupCoordinator: SecureBackupSetupCoordinatorType {
     
     // MARK: - Setup
     
-    init(session: MXSession, navigationRouter: NavigationRouterType? = nil) {
+    /// Initializer
+    /// - Parameters:
+    ///   - session: The MXSession.
+    ///   - checkKeyBackup: Indicate false to ignore existing key backup.
+    ///   - navigationRouter: Use existing navigation router to plug this flow or let nil to use new one.
+    init(session: MXSession, checkKeyBackup: Bool = true, navigationRouter: NavigationRouterType? = nil) {
         self.session = session
         self.recoveryService = session.crypto.recoveryService
         self.keyBackup = session.crypto.backup
+        self.checkKeyBackup = checkKeyBackup
         
         if let navigationRouter = navigationRouter {
             self.navigationRouter = navigationRouter
@@ -70,14 +77,15 @@ final class SecureBackupSetupCoordinator: SecureBackupSetupCoordinatorType {
     // MARK: - Private methods
 
     private func createIntro() -> SecureBackupSetupIntroViewController {
-        let introViewController = SecureBackupSetupIntroViewController.instantiate()
+        // TODO: Use a coordinator
+        let viewModel = SecureBackupSetupIntroViewModel(keyBackup: self.keyBackup, checkKeyBackup: self.checkKeyBackup)
+        let introViewController = SecureBackupSetupIntroViewController.instantiate(with: viewModel)
         introViewController.delegate = self
-        introViewController.keyBackup = self.keyBackup
         return introViewController
     }
     
-    private func showSetupKey(passphrase: String? = nil) {
-        let coordinator = SecretsSetupRecoveryKeyCoordinator(recoveryService: self.recoveryService, passphrase: passphrase)
+    private func showSetupKey(passphraseOnly: Bool, passphrase: String? = nil) {
+        let coordinator = SecretsSetupRecoveryKeyCoordinator(recoveryService: self.recoveryService, passphrase: passphrase, passphraseOnly: passphraseOnly)
         coordinator.delegate = self
         coordinator.start()
         
@@ -153,7 +161,7 @@ final class SecureBackupSetupCoordinator: SecureBackupSetupCoordinatorType {
 extension SecureBackupSetupCoordinator: SecureBackupSetupIntroViewControllerDelegate {
     
     func secureBackupSetupIntroViewControllerDidTapUseKey(_ secureBackupSetupIntroViewController: SecureBackupSetupIntroViewController) {
-        self.showSetupKey()
+        self.showSetupKey(passphraseOnly: false)
     }
     
     func secureBackupSetupIntroViewControllerDidTapUsePassphrase(_ secureBackupSetupIntroViewController: SecureBackupSetupIntroViewController) {
@@ -193,7 +201,7 @@ extension SecureBackupSetupCoordinator: SecretsSetupRecoveryPassphraseCoordinato
     }
     
     func secretsSetupRecoveryPassphraseCoordinator(_ coordinator: SecretsSetupRecoveryPassphraseCoordinatorType, didConfirmPassphrase passphrase: String) {
-        self.showSetupKey(passphrase: passphrase)
+        self.showSetupKey(passphraseOnly: false, passphrase: passphrase)        
     }
     
     func secretsSetupRecoveryPassphraseCoordinatorDidCancel(_ coordinator: SecretsSetupRecoveryPassphraseCoordinatorType) {